# gen/openapiv2/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "sdl/v1/models/models.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/canvases": {
      "get": {
        "summary": "List all canvases from a user.",
        "operationId": "CanvasService_ListCanvases",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListCanvasesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.pageKey",
            "description": "Instead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.pageOffset",
            "description": "If a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "Number of results to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "CanvasService"
        ]
      },
      "post": {
        "summary": "Create a new canvas sesssion.",
        "operationId": "CanvasService_CreateCanvas",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateCanvasResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateCanvasRequest"
            }
          }
        ],
        "tags": [
          "CanvasService"
        ]
      }
    },
...
    "/v1/systems": {
      "get": {
        "summary": "ListSystems returns all available systems",
        "operationId": "SystemsService_ListSystems",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListSystemsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "SystemsService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1Canvas": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string",
          "title": "Unique ID for the canvas"
        },
        "name": {
          "type": "string",
          "title": "Human-readable name for the canvas"
        },
        "description": {
          "type": "string",
          "title": "Description of what this canvas is for"
        },
        "activeSystem": {
          "type": "string",
          "title": "The currently active system (from the systems defined in system_contents)"
        },
        "systemContents": {
          "type": "string",
          "title": "Contents of the .sdl file that defines one or more systems"
        },
        "recipes": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Recipe files for various scenarios (name -\u003e contents map)"
        },
        "generators": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Generator"
          },
          "title": "Registered generators for this canvas"
        },
        "metrics": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Metric"
          },
          "title": "Registered live metrics for this canvas"
        }
      }
    },
...
    "v1CreateCanvasRequest": {
      "type": "object",
      "properties": {
        "canvas": {
          "$ref": "#/definitions/v1Canvas"
        }
      }
    },
    "v1CreateCanvasResponse": {
      "type": "object",
      "properties": {
        "canvas": {
          "$ref": "#/definitions/v1Canvas"
        },
        "fieldErrors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
    "v1Generator": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string",
          "title": "ID of the generator"
        },
        "canvasId": {
          "type": "string",
          "title": "Canvas this generator is sending traffic to"
        },
        "name": {
          "type": "string",
          "title": "A descriptive label"
        },
        "component": {
          "type": "string",
          "title": "Name of the target component to generate traffic on. This component should be defined in the System,\neg \"server\""
        },
        "method": {
          "type": "string",
          "description": "Method in the target component to generate traffic on."
        },
        "rate": {
          "type": "number",
          "format": "double",
          "title": "Traffic rate in RPS (\u003e= 1).  Does not support \u003c 1 yet"
        },
        "duration": {
          "type": "number",
          "format": "double",
          "title": "Duration in seconds over which the genarator is run. 0 =\u003e For ever"
        },
        "enabled": {
          "type": "boolean",
          "title": "whether it is enabled or not"
        }
      }
    },
...
    "v1ListCanvasesResponse": {
      "type": "object",
      "properties": {
        "canvases": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Canvas"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationResponse"
        }
      }
    },
...
    "v1ListSystemsResponse": {
      "type": "object",
      "properties": {
        "systems": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1SystemInfo"
          }
        }
      }
    },
...
    "v1Metric": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string"
        },
        "canvasId": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "title": "A descriptive label"
        },
        "component": {
          "type": "string",
          "title": "Name of the target component to monitor\neg \"server\""
        },
        "methods": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "Method in the target component to generate traffic on."
        },
        "enabled": {
          "type": "boolean",
          "title": "whether it is enabled or not"
        },
        "metricType": {
          "type": "string",
          "title": "Type of metric \"count\" or \"latency\""
        },
        "aggregation": {
          "type": "string",
          "title": "Type of aggregation on the metric"
        },
        "aggregationWindow": {
          "type": "number",
          "format": "double",
          "title": "Aggregation window (in seconds) to match on"
        },
        "matchResult": {
          "type": "string",
          "title": "Result value to match"
        },
        "matchResultType": {
          "type": "string",
          "title": "The result \"type\" if a matching result is provided\nThis will be parsed into a type declaration so we know how to treat\nthe match_result value provided"
        },
        "oldestTimestamp": {
          "type": "number",
          "format": "double"
        },
        "newestTimestamp": {
          "type": "number",
          "format": "double"
        },
        "numDataPoints": {
          "type": "string",
          "format": "int64",
          "title": "Number of data points stored"
        }
      }
    },
...
    "v1PaginationResponse": {
      "type": "object",
      "properties": {
        "nextPageKey": {
          "type": "string",
          "description": "The key/pointer string that subsequent List requests should pass to\ncontinue the pagination."
        },
        "nextPageOffset": {
          "type": "integer",
          "format": "int32",
          "title": "Also support an integer offset if possible"
        },
        "hasMore": {
          "type": "boolean",
          "description": "Whether theere are more results."
        },
        "totalResults": {
          "type": "integer",
          "format": "int32",
          "description": "Total number of results."
        }
      }
    },
...
    "v1SystemInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "difficulty": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "icon": {
          "type": "string"
        },
        "lastUpdated": {
          "type": "string"
        }
      },
      "title": "SystemInfo represents a system in the catalog"
    },
...
  }
}